Remove all references to my name
