+++ .gitignore
@@ -12,4 +12,4 @@
 
 .idea/
 
-/checksums/michael_taylor_sdk-1.0.0.gem.sha512
+/checksums/
