+++ lib/elephant_in_the_room/the_one_api_sdk/constants.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module ElephantInTheRoom::TheOneApiSdk
+  module Constants
+    MOVIE_PATH = "movie"
+    QUOTE_PATH = "quote"
+
+    EXPONENTIAL_BACKOFF_TIME_DIVISOR = 20
+  end
+end
