+++ lib/elephant_in_the_room/the_one_api_sdk/retry_strategy/one_try.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module ElephantInTheRoom::TheOneApiSdk::RetryStrategy
+  ##
+  # The default retry strategy that just runs a request once and passes along any errors
+  class OneTry
+    def run(do_request)
+      do_request.call
+    end
+  end
+end
