+++ lib/elephant_in_the_room/the_one_api_sdk/version.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module ElephantInTheRoom
+  module TheOneApiSdk
+    VERSION = "2.0.0"
+  end
+end
