--- lib/michael_taylor_sdk/constants.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module MichaelTaylorSdk
-  module Constants
-    MOVIE_PATH = "movie"
-    QUOTE_PATH = "quote"
-
-    EXPONENTIAL_BACKOFF_TIME_DIVISOR = 20
-  end
-end
