--- lib/michael_taylor_sdk/retry_strategy/one_try.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module MichaelTaylorSdk::RetryStrategy
-  ##
-  # The default retry strategy that just runs a request once and passes along any errors
-  class OneTry
-    def run(do_request)
-      do_request.call
-    end
-  end
-end
