--- lib/michael_taylor_sdk/version.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-module MichaelTaylorSdk
-  VERSION = "2.0.0"
-end
