+++ sig/michael_taylor_sdk.rbs
@@ -1,4 +1,4 @@
-module MichaelTaylorSdk
+module ElephantInTheRoom::TheOneApiSdk
   VERSION: String
   # See the writing guide of rbs: https://github.com/ruby/rbs#guides
 end
