+++ spec/elephant_in_the_room/the_one_api_sdk_spec.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+RSpec.describe ElephantInTheRoom::TheOneApiSdk do
+  it "has a version number" do
+    expect(ElephantInTheRoom::TheOneApiSdk::VERSION).not_to be_nil
+  end
+end
