--- spec/michael_taylor_sdk_spec.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-RSpec.describe MichaelTaylorSdk do
-  it "has a version number" do
-    expect(MichaelTaylorSdk::VERSION).not_to be_nil
-  end
-end
